+++ demos/client-example-server/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @prosopo/client-example-server
 
+## 2.7.89
+### Patch Changes
+
+- Updated dependencies [6c26669]
+- Updated dependencies [f7f9ec5]
+  - @prosopo/types@4.2.1
+  - @prosopo/keyring@2.9.33
+  - @prosopo/server@2.10.11
+
 ## 2.7.88
 ### Patch Changes