+++ packages/datasets/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @prosopo/datasets
 
+## 3.1.27
+### Patch Changes
+
+- Updated dependencies [6c26669]
+- Updated dependencies [f7f9ec5]
+  - @prosopo/types@4.2.1
+
 ## 3.1.26
 ### Patch Changes