+++ packages/procaptcha-wrapper/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @prosopo/procaptcha-wrapper
 
+## 2.6.77
+### Patch Changes
+
+- Updated dependencies [6c26669]
+- Updated dependencies [f7f9ec5]
+  - @prosopo/types@4.2.1
+
 ## 2.6.76
 ### Patch Changes