+++ packages/user-access-policy/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @prosopo/user-access-policy
 
+## 3.7.10
+### Patch Changes
+
+- Updated dependencies [6c26669]
+- Updated dependencies [f7f9ec5]
+  - @prosopo/types@4.2.1
+  - @prosopo/api@3.4.7
+
 ## 3.7.9
 ### Patch Changes